Add sisrs.py file
